python_packages/shabti_keycloak/pyproject.toml
[project]
name = "shabti-keycloak"
version = "0.1a2"
requires-python = ">=3.12"
dependencies = [
  "python-keycloak"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"